Move deep_hash_accessor.rb to lib/seam
