+++ lib/seam/deep_hash_accessor.rb
